backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "srishti"
version = "0.1.0"
description = "AI-Powered Multi-Agent Event Intelligence Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-pinecone>=0.2.0",
    "langgraph>=0.2.0",
    "groq>=0.11.0",
    "pinecone-client>=5.0.0",
    "sentence-transformers>=3.0.0",
    "supabase>=2.9.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "playwright>=1.47.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]
